refactor: restructure screens
