+++ packages/cryptoplease/lib/app/app.dart
@@ -2,7 +2,6 @@ import 'package:auto_route/auto_route.dart';
 import 'package:cryptoplease/app/routes.dart';
 import 'package:cryptoplease/core/accounts/bl/accounts_bloc.dart';
 import 'package:cryptoplease/features/app_lock/module.dart';
-import 'package:cryptoplease/features/incoming_split_key_payment/bl/bloc.dart';
 import 'package:cryptoplease_ui/cryptoplease_ui.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -25,9 +24,6 @@ class _CryptopleaseAppState extends State<CryptopleaseApp> {
         context.select<AccountsBloc, bool>((b) => b.state.isProcessing);
     final isAuthenticated =
         context.select<AccountsBloc, bool>((b) => b.state.account != null);
-    final skState = context.watch<SplitKeyIncomingPaymentBloc>().state;
-    final shouldLoginForSk =
-        !isAuthenticated && skState is PaymentSecondPartReady;
 
     return CpTheme(
       theme: const CpThemeData.light(),
@@ -43,8 +39,6 @@ class _CryptopleaseAppState extends State<CryptopleaseApp> {
                 const SplashRoute()
               else
                 const SignUpFlowRoute(),
-              if (skState is! PaymentNone && !shouldLoginForSk)
-                const SplitKeyIncomingPaymentRoute()
             ],
           ),
           localizationsDelegates: AppLocalizations.localizationsDelegates,
