+++ packages/cryptoplease/lib/app/screens/authenticated/routes.dart
@@ -1,8 +1,9 @@
 import 'package:auto_route/auto_route.dart';
-import 'package:cryptoplease/app/screens/authenticated/flow.dart';
-import 'package:cryptoplease/app/screens/authenticated/home_tabs_screen.dart';
+import 'package:cryptoplease/app/screens/authenticated/authenticated_flow_screen.dart';
+import 'package:cryptoplease/app/screens/authenticated/home_screen.dart';
+import 'package:cryptoplease/app/screens/authenticated/investments_screen.dart';
 import 'package:cryptoplease/app/screens/authenticated/receive_flow/routes.dart';
-import 'package:cryptoplease/app/screens/authenticated/wallet_screen.dart';
+import 'package:cryptoplease/app/screens/authenticated/wallet_flow/wallet_flow_screen.dart';
 import 'package:cryptoplease/core/presentation/token_selector_screen.dart';
 import 'package:cryptoplease/core/tokens/token.dart';
 import 'package:cryptoplease/features/activities/presentation/activities_screen.dart';
@@ -20,8 +21,6 @@ import 'package:cryptoplease/features/profile/edit_profile_screen.dart';
 import 'package:cryptoplease/features/profile/help_screen.dart';
 import 'package:cryptoplease/features/profile/menu_screen.dart';
 import 'package:cryptoplease/features/profile/profile_screen.dart';
-import 'package:cryptoplease/features/request_pay/presentation/request_pay_flow.dart';
-import 'package:cryptoplease/features/request_pay/presentation/routes.dart';
 import 'package:cryptoplease/features/swap_tokens/presentation/routes.dart';
 import 'package:cryptoplease/features/swap_tokens/presentation/swap_token_flow.dart';
 import 'package:cryptoplease/features/transactions/presentation/transaction_details_screen/transaction_details_screen.dart';
@@ -32,15 +31,14 @@ const authenticatedFlowRoutes = AutoRoute<dynamic>(
   children: [
     AutoRoute<void>(
       path: '',
-      page: HomeTabsScreen,
+      page: HomeScreen,
       children: [
         CustomRoute<void>(path: '', page: MenuScreen),
-        CustomRoute<void>(page: RequestPayFlowScreen),
+        CustomRoute<void>(page: WalletFlowScreen),
+        CustomRoute<void>(page: InvestmentsScreen),
         CustomRoute<void>(page: ActivitiesScreen),
-        CustomRoute<void>(page: WalletScreen),
       ],
     ),
-    ...requestDirectRoutes,
     ...backupPhraseRoutes,
     ...odpRoutes,
     ...oskpRoutes,
