+++ packages/cryptoplease/lib/features/incoming_split_key_payments/presentation/first_part_ready_screen.dart
@@ -1,6 +1,6 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:cryptoplease/app/routes.dart';
-import 'package:cryptoplease/app/screens/dynamic_links/dynamic_links_notifier.dart';
+import 'package:cryptoplease/core/dynamic_links_notifier.dart';
 import 'package:cryptoplease/features/incoming_split_key_payment/bl/tx_processor.dart';
 import 'package:cryptoplease/features/incoming_split_key_payments/bl/iskp_bloc.dart';
 import 'package:cryptoplease/features/incoming_split_key_payments/bl/pending_iskp.dart';
