+++ packages/cryptoplease/lib/features/outgoing_split_key_payments/presentation/routes.dart
@@ -1,8 +1,10 @@
 import 'package:auto_route/auto_route.dart';
+import 'package:cryptoplease/features/outgoing_split_key_payments/presentation/oskp_confirmation_screen.dart';
 import 'package:cryptoplease/features/outgoing_split_key_payments/presentation/oskp_screen.dart';
 import 'package:cryptoplease/features/outgoing_split_key_payments/presentation/share_links_screen.dart';
 
 const oskpRoutes = [
   AutoRoute<void>(page: OSKPScreen),
+  AutoRoute<void>(page: OSKPConfirmationScreen),
   AutoRoute<void>(page: ShareLinksScreen),
 ];
